# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarTrack",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Default server"
        }
    ],
    "paths": {
        "/api/affiliations/create": {
            "post": {
                "tags": [
                    "Affiliations"
                ],
                "summary": "create affiliations",
                "operationId": "App\\Http\\Controllers\\AffiliationController::create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "hero_id": "2",
                                    "name": "superman"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/affiliations/{id}/remove": {
            "delete": {
                "tags": [
                    "Affiliations"
                ],
                "summary": "remove one affiliation with id",
                "operationId": "App\\Http\\Controllers\\AffiliationController::remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register a new user and send notification mail",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirm password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login an user",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Wrong combination of email and password or email not verified"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout an user",
                "operationId": "logout",
                "responses": {
                    "204": {
                        "description": "Successful operation with no content in return"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/getUser": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get user",
                "description": "Retrieve information from current user",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/password/change": {
            "patch": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change password",
                "description": "Change an user's password (requires current password) and send notification mail",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "new_password": {
                                        "description": "New password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "new_password_confirmation": {
                                        "description": "Confirm new password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Wrong combination of email and password or email not verified"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    }
}